log USE_ON_DISK_DKT env var
